- moving sqlite to sqlalchemy login: auth_sqlalchemy.py, db_models.py - renew: requirements.txt, README.md, main.py, gui_login.py - removed: login_db.py
