+++ src/db_models.py
@@ -1,0 +1,15 @@
+from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy.orm import declarative_base, sessionmaker
+
+Base = declarative_base()
+
+class User(Base):
+    __tablename__ = 'users'
+
+    id = Column(Integer, primary_key=True)
+    username = Column(String, unique=True, nullable=False)
+    password = Column(String, nullable=False)
+
+# Create engine and session
+engine = create_engine("sqlite:///finance_tracker.db", echo=True)
+Session = sessionmaker(bind=engine)
